cli/requirements.txt
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
scipy==1.11.1
